Remove default export from usePersistAppData.ts
You can have only one default export in a module,
whereas you can have as many named exports as you like.

If you provide a default export, the programmer
using it has to come up with a name for it.
This can lead to inconsistency in a codebase.

In contrast, with a named export, the programmer
doesn't have to think about what to call it unless
there's a conflict with another identifier in their module.

With a named export, the person importing it has
to specify the name of what they're importing.
They get a nice early error if they try to import
something that doesn't exist.

If you consistently only use named exports,
programmers importing from modules in the project
don't have to think about whether what they want
is the default or a named export.

ref. https://stackoverflow.com/questions/46913851/why-and-when-to-use-default-export-over-named-exports-in-es6-modules